Search Google Scholar Highly Related Papers or Author

Search Google Scholar Highly Related Papers or Author

scholar_search(
  keyword,
  is_author = TRUE,
  server_url = "https://api.scaleserp.com",
  server_key = NULL
)

Arguments

keyword

A keyword, can be author name, e.g. "Shixiang Wang".

is_author

Default is TRUE, find author information, if FALSE, return the first page result in search engine.

server_url

Server URL, here I use Scale SERP API.

server_key

Key for searching data (multiple is acceptable), you can obtain it from URL above. If not set, use personal key from Shixiang. Total 125x2 free searches per month.

Value

A data.frame or a list.

Examples

# \donttest{
x <- scholar_search("Shixiang Wang")
#> Searching author Shixiang Wang
#> Using API server: https://api.scaleserp.com
#> Using Shixiang's personal API key, only 125x2 free searches per month for all packages users
#> querying with the 1th API key
#> Search times used: 4
#> Search times left: 121
x
#>             id
#> 1 FvNp0NkAAAAJ
#> 2 JT5dCdYAAAAJ
#>                                                                        desc
#> 1   Shixiang Wang (王诗翔) - Verified email at sysucc.org.cn - Cited by 509
#> 2 Shixiang Wang (王施相) - Verified email at connect.polyu.hk - Cited by 73
x <- scholar_search("Shixiang Wang", is_author = FALSE)
#> Searching Shixiang Wang
#> Using API server: https://api.scaleserp.com
#> Using Shixiang's personal API key, only 125x2 free searches per month for all packages users
#> querying with the 1th API key
#> Search times used: 5
#> Search times left: 120
if (!is.null(x)) {
  x$gt
}
#> <div id="agxmkrmelz" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #agxmkrmelz .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #agxmkrmelz .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #agxmkrmelz .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #agxmkrmelz .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #agxmkrmelz .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #agxmkrmelz .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #agxmkrmelz .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #agxmkrmelz .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #agxmkrmelz .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #agxmkrmelz .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #agxmkrmelz .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #agxmkrmelz .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #agxmkrmelz .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #agxmkrmelz .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #agxmkrmelz .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #agxmkrmelz .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #agxmkrmelz .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #agxmkrmelz .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #agxmkrmelz .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #agxmkrmelz .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #agxmkrmelz .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #agxmkrmelz .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #agxmkrmelz .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #agxmkrmelz .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #agxmkrmelz .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #agxmkrmelz .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #agxmkrmelz .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #agxmkrmelz .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #agxmkrmelz .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-left: 4px;
#>   padding-right: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #agxmkrmelz .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #agxmkrmelz .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #agxmkrmelz .gt_left {
#>   text-align: left;
#> }
#> 
#> #agxmkrmelz .gt_center {
#>   text-align: center;
#> }
#> 
#> #agxmkrmelz .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #agxmkrmelz .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #agxmkrmelz .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #agxmkrmelz .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #agxmkrmelz .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #agxmkrmelz .gt_two_val_uncert {
#>   display: inline-block;
#>   line-height: 1em;
#>   text-align: right;
#>   font-size: 60%;
#>   vertical-align: -0.25em;
#>   margin-left: 0.1em;
#> }
#> 
#> #agxmkrmelz .gt_footnote_marks {
#>   font-style: italic;
#>   font-weight: normal;
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#> }
#> 
#> #agxmkrmelz .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #agxmkrmelz .gt_slash_mark {
#>   font-size: 0.7em;
#>   line-height: 0.7em;
#>   vertical-align: 0.15em;
#> }
#> 
#> #agxmkrmelz .gt_fraction_numerator {
#>   font-size: 0.6em;
#>   line-height: 0.6em;
#>   vertical-align: 0.45em;
#> }
#> 
#> #agxmkrmelz .gt_fraction_denominator {
#>   font-size: 0.6em;
#>   line-height: 0.6em;
#>   vertical-align: -0.05em;
#> }
#> </style>
#>   <table class="gt_table">
#>   
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">title</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">link</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">domain</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">displayed_link</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">snippet</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">cited_by_count</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">cited_by_link</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">versions_count</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">versions_link</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td class="gt_row gt_left">Antigen presentation and tumor immunogenicity in cancer immunotherapy response prediction</td>
#> <td class="gt_row gt_left">https://elifesciences.org/articles/49020</td>
#> <td class="gt_row gt_left">elifesciences.org</td>
#> <td class="gt_row gt_left">S Wang, Z He, X Wang, H Li, XS Liu - Elife, 2019 - elifesciences.org</td>
#> <td class="gt_row gt_left">10.7554/eLife.49020.001 Immunotherapy, represented by immune checkpoint inhibitors (ICI), 
#> is transforming the treatment of cancer. However, only a small percentage of patients show …</td>
#> <td class="gt_row gt_right">133</td>
#> <td class="gt_row gt_left">https://www.google.com/scholar?cites=7504395953338341860&amp;as_sdt=2005&amp;sciodt=0,5&amp;hl=en</td>
#> <td class="gt_row gt_right">10</td>
#> <td class="gt_row gt_left">https://www.google.com/scholar?cluster=7504395953338341860&amp;hl=en&amp;as_sdt=0,5</td></tr>
#>     <tr><td class="gt_row gt_left">[HTML][HTML] APOBEC3B and APOBEC mutational signature as potential predictive markers for immunotherapy response in non-small cell lung cancer</td>
#> <td class="gt_row gt_left">https://www.nature.com/articles/s41388-018-0245-9</td>
#> <td class="gt_row gt_left">www.nature.com</td>
#> <td class="gt_row gt_left">S Wang, M Jia, Z He, XS Liu - Oncogene, 2018 - nature.com</td>
#> <td class="gt_row gt_left">Non-small cell lung cancer (NSCLC) is known to carry heavy mutation load. Besides 
#> smoking, cytidine deaminase APOBEC3B plays a key role in the mutation process of NSCLC. …</td>
#> <td class="gt_row gt_right">144</td>
#> <td class="gt_row gt_left">https://www.google.com/scholar?cites=7574700313752091686&amp;as_sdt=2005&amp;sciodt=0,5&amp;hl=en</td>
#> <td class="gt_row gt_right">8</td>
#> <td class="gt_row gt_left">https://www.google.com/scholar?cluster=7574700313752091686&amp;hl=en&amp;as_sdt=0,5</td></tr>
#>     <tr><td class="gt_row gt_left">Femtosecond laser direct writing of flexibly configured waveguide geometries in optical crystals: fabrication and application</td>
#> <td class="gt_row gt_left">https://www.oejournal.org/article/doi/10.29026/oea.2020.190042?viewType=HTML</td>
#> <td class="gt_row gt_left">www.oejournal.org</td>
#> <td class="gt_row gt_left">Y Jia, S Wang, F Chen - Opto-Electronic Advances, 2020 - oejournal.org</td>
#> <td class="gt_row gt_left">Optical waveguides are far more than mere connecting elements in integrated optical systems 
#> and circuits. Benefiting from their high optical confinement and miniaturized footprints, …</td>
#> <td class="gt_row gt_right">71</td>
#> <td class="gt_row gt_left">https://www.google.com/scholar?cites=15248006345193113203&amp;as_sdt=2005&amp;sciodt=0,5&amp;hl=en</td>
#> <td class="gt_row gt_right">8</td>
#> <td class="gt_row gt_left">https://www.google.com/scholar?cluster=15248006345193113203&amp;hl=en&amp;as_sdt=0,5</td></tr>
#>     <tr><td class="gt_row gt_left">Sex differences in cancer immunotherapy efficacy, biomarkers, and therapeutic strategy</td>
#> <td class="gt_row gt_left">https://www.mdpi.com/527894</td>
#> <td class="gt_row gt_left">www.mdpi.com</td>
#> <td class="gt_row gt_left">S Wang, LA Cowley, XS Liu - Molecules, 2019 - mdpi.com</td>
#> <td class="gt_row gt_left">Sex differences in innate and adaptive immune responses are known, and women generally 
#> mount a stronger immune response than men. Cancer immunotherapy, represented by …</td>
#> <td class="gt_row gt_right">84</td>
#> <td class="gt_row gt_left">https://www.google.com/scholar?cites=2863414971057974891&amp;as_sdt=2005&amp;sciodt=0,5&amp;hl=en</td>
#> <td class="gt_row gt_right">11</td>
#> <td class="gt_row gt_left">https://www.google.com/scholar?cluster=2863414971057974891&amp;hl=en&amp;as_sdt=0,5</td></tr>
#>     <tr><td class="gt_row gt_left">[HTML][HTML] A novel compound DBZ ameliorates neuroinflammation in LPS-stimulated microglia and ischemic stroke rats: Role of Akt (Ser473)/GSK3β (Ser9)-mediated …</td>
#> <td class="gt_row gt_left">https://www.sciencedirect.com/science/article/pii/S2213231720308491</td>
#> <td class="gt_row gt_left">www.sciencedirect.com</td>
#> <td class="gt_row gt_left">S Liao, J Wu, R Liu, S Wang, J Luo, Y Yang, Y Qin, T Li… - Redox biology, 2020 - Elsevier</td>
#> <td class="gt_row gt_left">Microglia-mediated neuroinflammation plays a crucial role in the pathophysiological process 
#> of multiple neurological disorders such as ischemic stroke, yet lacks effective therapeutic …</td>
#> <td class="gt_row gt_right">42</td>
#> <td class="gt_row gt_left">https://www.google.com/scholar?cites=3905713535682111782&amp;as_sdt=2005&amp;sciodt=0,5&amp;hl=en</td>
#> <td class="gt_row gt_right">6</td>
#> <td class="gt_row gt_left">https://www.google.com/scholar?cluster=3905713535682111782&amp;hl=en&amp;as_sdt=0,5</td></tr>
#>     <tr><td class="gt_row gt_left">Effective degradation of sulfamethoxazole with Fe2+-zeolite/peracetic acid</td>
#> <td class="gt_row gt_left">https://www.sciencedirect.com/science/article/pii/S1383586619318593</td>
#> <td class="gt_row gt_left">www.sciencedirect.com</td>
#> <td class="gt_row gt_left">S Wang, H Wang, Y Liu, Y Fu - Separation and Purification Technology, 2020 - Elsevier</td>
#> <td class="gt_row gt_left">Peracetic acid (PAA), a green disinfectant, has been widely used in the wastewater treatment. 
#> Here, a novel heterogeneous advanced oxidation process (Fe 2+ -zeolite/PAA) capable of …</td>
#> <td class="gt_row gt_right">52</td>
#> <td class="gt_row gt_left">https://www.google.com/scholar?cites=15046443596922888471&amp;as_sdt=2005&amp;sciodt=0,5&amp;hl=en</td>
#> <td class="gt_row gt_right">NA</td>
#> <td class="gt_row gt_left">NA</td></tr>
#>     <tr><td class="gt_row gt_left">HCO3–/CO32–enhanced degradation of diclofenac by Cu (Ⅱ)-activated peracetic acid: Efficiency and mechanism</td>
#> <td class="gt_row gt_left">https://www.sciencedirect.com/science/article/pii/S1383586621011424</td>
#> <td class="gt_row gt_left">www.sciencedirect.com</td>
#> <td class="gt_row gt_left">Z Wang, Y Fu, Y Peng, S Wang, Y Liu - Separation and Purification …, 2021 - Elsevier</td>
#> <td class="gt_row gt_left">Diclofenac (DCF), a non-steroidal anti-inflammatory drug, is an emerging and nonbiodegradable 
#> contaminant which is detected extensively in aquatic environment. Therefore, low cost …</td>
#> <td class="gt_row gt_right">13</td>
#> <td class="gt_row gt_left">https://www.google.com/scholar?cites=58391578547929123&amp;as_sdt=2005&amp;sciodt=0,5&amp;hl=en</td>
#> <td class="gt_row gt_right">2</td>
#> <td class="gt_row gt_left">https://www.google.com/scholar?cluster=58391578547929123&amp;hl=en&amp;as_sdt=0,5</td></tr>
#>     <tr><td class="gt_row gt_left">Strong and tough self-healing elastomers enabled by dual reversible networks formed by ionic interactions and dynamic covalent bonds</td>
#> <td class="gt_row gt_left">https://www.sciencedirect.com/science/article/pii/S0032386118308802</td>
#> <td class="gt_row gt_left">www.sciencedirect.com</td>
#> <td class="gt_row gt_left">Y Peng, Y Yang, Q Wu, S Wang, G Huang, J Wu - Polymer, 2018 - Elsevier</td>
#> <td class="gt_row gt_left">A strong and tough self-healing elastomer is prepared based on double reversible networks 
#> consisting of ionic interactions and Diels-Alder (DA) crosslinks. The elastomer is …</td>
#> <td class="gt_row gt_right">60</td>
#> <td class="gt_row gt_left">https://www.google.com/scholar?cites=6814477063818521582&amp;as_sdt=2005&amp;sciodt=0,5&amp;hl=en</td>
#> <td class="gt_row gt_right">2</td>
#> <td class="gt_row gt_left">https://www.google.com/scholar?cluster=6814477063818521582&amp;hl=en&amp;as_sdt=0,5</td></tr>
#>     <tr><td class="gt_row gt_left">Degradation of diclofenac by Fe (II)-activated bisulfite: Kinetics, mechanism and transformation products</td>
#> <td class="gt_row gt_left">https://www.sciencedirect.com/science/article/pii/S0045653519317424</td>
#> <td class="gt_row gt_left">www.sciencedirect.com</td>
#> <td class="gt_row gt_left">H Wang, S Wang, Y Liu, Y Fu, P Wu, G Zhou - Chemosphere, 2019 - Elsevier</td>
#> <td class="gt_row gt_left">As an emerging pollutant, Diclofenac (DCF) has potential threats to ecosystem and human 
#> health, and it can hardly be removed by conventional wastewater treatment processes. In this …</td>
#> <td class="gt_row gt_right">40</td>
#> <td class="gt_row gt_left">https://www.google.com/scholar?cites=16624056245035934411&amp;as_sdt=2005&amp;sciodt=0,5&amp;hl=en</td>
#> <td class="gt_row gt_right">7</td>
#> <td class="gt_row gt_left">https://www.google.com/scholar?cluster=16624056245035934411&amp;hl=en&amp;as_sdt=0,5</td></tr>
#>     <tr><td class="gt_row gt_left">UCSCXenaShiny: an R/CRAN package for interactive analysis of UCSC Xena data</td>
#> <td class="gt_row gt_left">https://academic.oup.com/bioinformatics/article-abstract/38/2/527/6330614</td>
#> <td class="gt_row gt_left">academic.oup.com</td>
#> <td class="gt_row gt_left">…, K Gu, Y Li, F Zhao, J Li, M Wang, H Wang… - …, 2022 - academic.oup.com</td>
#> <td class="gt_row gt_left">UCSC Xena platform provides huge amounts of processed cancer omics data from large 
#> cancer research projects (eg TCGA, CCLE and PCAWG) or individual research groups and …</td>
#> <td class="gt_row gt_right">18</td>
#> <td class="gt_row gt_left">https://www.google.com/scholar?cites=5372423792264139063&amp;as_sdt=2005&amp;sciodt=0,5&amp;hl=en</td>
#> <td class="gt_row gt_right">8</td>
#> <td class="gt_row gt_left">https://www.google.com/scholar?cluster=5372423792264139063&amp;hl=en&amp;as_sdt=0,5</td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>
# }