Search Google Scholar Highly Related Papers or Author

Search Google Scholar Highly Related Papers or Author

scholar_search(
  keyword,
  is_author = TRUE,
  server_url = "https://api.scaleserp.com",
  server_key = NULL
)

Arguments

keyword
A keyword, can be author name, e.g. "Shixiang Wang".

is_author
Default is TRUE, find author information, if FALSE, return the first page result in search engine.

server_url
Server URL, here I use Scale SERP API.

server_key
Key for searching data, you can obtain it from URL above. If not set, use personal key from Shixiang. Total 125 free searches per month.

Value

A data.frame or a list.

Examples

# \donttest{
x <- scholar_search("Shixiang Wang")

#> Searching author Shixiang Wang

#> Using API server: https://api.scaleserp.com

#> Using Shixiang's personal API key, only 125 free searches per month for all packages users

#> Search times used: 28

#> Search times left: 97

x

#>             id
#> 1 FvNp0NkAAAAJ
#>                                                                                               desc
#> 1 Wang, Shixiang (王诗翔)ShanghaiTech. UniversityVerified email at shanghaitech.edu.cnCited by 119

x <- scholar_search("Shixiang Wang", is_author = FALSE)

#> Searching Shixiang Wang

#> Using API server: https://api.scaleserp.com

#> Using Shixiang's personal API key, only 125 free searches per month for all packages users

#> Search times used: 29

#> Search times left: 96

x$gt

#> <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #mwlefoodfi .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #mwlefoodfi .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #mwlefoodfi .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #mwlefoodfi .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 4px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #mwlefoodfi .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #mwlefoodfi .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #mwlefoodfi .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #mwlefoodfi .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #mwlefoodfi .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #mwlefoodfi .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #mwlefoodfi .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #mwlefoodfi .gt_group_heading {
#>   padding: 8px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #mwlefoodfi .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #mwlefoodfi .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #mwlefoodfi .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #mwlefoodfi .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #mwlefoodfi .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 12px;
#> }
#> 
#> #mwlefoodfi .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #mwlefoodfi .gt_first_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #mwlefoodfi .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #mwlefoodfi .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #mwlefoodfi .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #mwlefoodfi .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #mwlefoodfi .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #mwlefoodfi .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #mwlefoodfi .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #mwlefoodfi .gt_sourcenote {
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #mwlefoodfi .gt_left {
#>   text-align: left;
#> }
#> 
#> #mwlefoodfi .gt_center {
#>   text-align: center;
#> }
#> 
#> #mwlefoodfi .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #mwlefoodfi .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #mwlefoodfi .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #mwlefoodfi .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #mwlefoodfi .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #mwlefoodfi .gt_footnote_marks {
#>   font-style: italic;
#>   font-size: 65%;
#> }
#> </style>
#> <div id="mwlefoodfi" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;"><table class="gt_table">
#>   
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">title</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">link</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">domain</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">displayed_link</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">snippet</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr>
#>       <td class="gt_row gt_left">[HTML][HTML] APOBEC3B and APOBEC mutational signature as potential predictive markers for immunotherapy response in non-small cell lung cancer</td>
#>       <td class="gt_row gt_left">https://www.nature.com/articles/s41388-018-0245-9</td>
#>       <td class="gt_row gt_left">www.nature.com</td>
#>       <td class="gt_row gt_left">S Wang, M Jia, Z He, XS Liu - Oncogene, 2018 - nature.com</td>
#>       <td class="gt_row gt_left">Non-small cell lung cancer (NSCLC) is known to carry heavy mutation load. Besides smoking, cytidine deaminase APOBEC3B plays a key role in the mutation process of NSCLC. APOBEC3B is also reported to be upregulated and predicts bad prognosis in …</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">[HTML][HTML] Antigen presentation and tumor immunogenicity in cancer immunotherapy response prediction</td>
#>       <td class="gt_row gt_left">https://elifesciences.org/articles/49020</td>
#>       <td class="gt_row gt_left">elifesciences.org</td>
#>       <td class="gt_row gt_left">S Wang, Z He, X Wang, H Li, XS Liu - Elife, 2019 - elifesciences.org</td>
#>       <td class="gt_row gt_left">Immunotherapy, represented by immune checkpoint inhibitors (ICI), is transforming the treatment of cancer. However, only a small percentage of patients show response to ICI, and there is an unmet need for biomarkers that will identify patients who are more likely to …</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">The predictive power of tumor mutational burden in lung cancer immunotherapy response is influenced by patients' sex</td>
#>       <td class="gt_row gt_left">https://onlinelibrary.wiley.com/doi/abs/10.1002/ijc.32327</td>
#>       <td class="gt_row gt_left">onlinelibrary.wiley.com</td>
#>       <td class="gt_row gt_left">S Wang, J Zhang, Z He, K Wu… - International journal of …, 2019 - Wiley Online Library</td>
#>       <td class="gt_row gt_left">Immunotherapy, represented by immune checkpoint inhibitors (ICI), is transforming the treatment of cancer. However, only a fraction of patients show response to ICI, and there is an unmet need for biomarkers that will identify patients more likely to respond to ICI. Here …</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">Strong and tough self-healing elastomers enabled by dual reversible networks formed by ionic interactions and dynamic covalent bonds</td>
#>       <td class="gt_row gt_left">https://www.sciencedirect.com/science/article/pii/S0032386118308802</td>
#>       <td class="gt_row gt_left">www.sciencedirect.com</td>
#>       <td class="gt_row gt_left">Y Peng, Y Yang, Q Wu, S Wang, G Huang, J Wu - Polymer, 2018 - Elsevier</td>
#>       <td class="gt_row gt_left">A strong and tough self-healing elastomer is prepared based on double reversible networks consisting of ionic interactions and Diels-Alder (DA) crosslinks. The elastomer is synthesized through one-pot copolymerization of a pair of oppositely charged monomers and furan …</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">Determination of the main bioactive metabolites of Radix Salvia miltiorrhizae in Compound Danshen Dripping Pills and the tissue distribution of Danshensu in rabbit by SPE …</td>
#>       <td class="gt_row gt_left">https://onlinelibrary.wiley.com/doi/abs/10.1002/jssc.200600287</td>
#>       <td class="gt_row gt_left">onlinelibrary.wiley.com</td>
#>       <td class="gt_row gt_left">Z Xiaohui, Z Xinfeng, Z Xin, W Shixiang… - Journal of separation …, 2007 - Wiley Online Library</td>
#>       <td class="gt_row gt_left">A fast solid phase extraction‐high performance liquid chromatography‐ion trap mass spectrometry (SPE‐HPLC‐MSn) method is presented for quantitative determination of 3‐(3, 4‐dihydroxyphenyl)‐2‐hydroxylpropionic acid (Danshensu, DSS) from Compound Danshen …</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">Mortality rates among prevalent hemodialysis patients in Beijing: a comparison with USRDS data</td>
#>       <td class="gt_row gt_left">https://academic.oup.com/ndt/article-abstract/28/3/724/1812690</td>
#>       <td class="gt_row gt_left">academic.oup.com</td>
#>       <td class="gt_row gt_left">…, T Cui, A Zhang, K Wang, S Wang… - Nephrology Dialysis …, 2013 - academic.oup.com</td>
#>       <td class="gt_row gt_left">Background The raw annual mortality rate reported in Chinese patients on maintenance hemodialysis (MHD) was around 10% between 2005 to 2010, and it was around 20% in the US as reported by the United States Renal Data System (USRDS). Our hypothesis was that …</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">Adsorptive removal of fluoride from aqueous solutions using Al-humic acid-La aerogel composites</td>
#>       <td class="gt_row gt_left">https://www.sciencedirect.com/science/article/pii/S1385894716309810</td>
#>       <td class="gt_row gt_left">www.sciencedirect.com</td>
#>       <td class="gt_row gt_left">Y Liu, Q Fan, S Wang, Y Liu, A Zhou, L Fan - Chemical Engineering …, 2016 - Elsevier</td>
#>       <td class="gt_row gt_left">A novel and environmental friendly aerogel composite was synthesized by freeze-drying method using humic acid as low cost starting material. The as-prepared aluminum-humic acid-lanthanum (Al-HA-La) aerogel composites were fully characterized by nitrogen …</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">Degradation of diclofenac by Fe (II)-activated bisulfite: Kinetics, mechanism and transformation products</td>
#>       <td class="gt_row gt_left">https://www.sciencedirect.com/science/article/pii/S0045653519317424</td>
#>       <td class="gt_row gt_left">www.sciencedirect.com</td>
#>       <td class="gt_row gt_left">H Wang, S Wang, Y Liu, Y Fu, P Wu, G Zhou - Chemosphere, 2019 - Elsevier</td>
#>       <td class="gt_row gt_left">As an emerging pollutant, Diclofenac (DCF) has potential threats to ecosystem and human health, and it can hardly be removed by conventional wastewater treatment processes. In this study, Fe (II)-activated bisulfite (BS), an advanced oxidation process, was used for rapid …</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">Thermodynamic study of the interaction between terbutaline and salbutamol with an immobilized β2-adrenoceptor by high-performance liquid chromatography</td>
#>       <td class="gt_row gt_left">https://www.sciencedirect.com/science/article/pii/S1570023209001196</td>
#>       <td class="gt_row gt_left">www.sciencedirect.com</td>
#>       <td class="gt_row gt_left">X Zhao, X Zheng, Y Wei, L Bian, S Wang… - … of Chromatography B, 2009 - Elsevier</td>
#>       <td class="gt_row gt_left">Investigation of the interaction between drugs and receptors is very important in revealing the biologic basis and mechanism of the drug, and designing new bioactive compounds. The β 2-adrenoceptor (β 2-AR) was purified from rabbit lung tissues and immobilized on the …</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">Serum magnesium and mortality in maintenance hemodialysis patients</td>
#>       <td class="gt_row gt_left">https://www.karger.com/Article/Abstract/451052</td>
#>       <td class="gt_row gt_left">www.karger.com</td>
#>       <td class="gt_row gt_left">L Yu, H Li, S Wang - Blood purification, 2017 - karger.com</td>
#>       <td class="gt_row gt_left">Background and Aim: The study aimed to investigate the potential contributing effect of serum magnesium on mortality in maintenance hemodialysis (MHD) patients. Methods: The patients receiving regular MHD in March 2013 were involved. Baseline data including …</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">Effects of ionic liquid and nanogold particles on high-performance liquid chromatography-electrochemical detection and their application in highly efficient separation …</td>
#>       <td class="gt_row gt_left">https://www.sciencedirect.com/science/article/pii/S003991401201065X</td>
#>       <td class="gt_row gt_left">www.sciencedirect.com</td>
#>       <td class="gt_row gt_left">P Jia, S Wang, X Meng, W Lan, J Luo, S Liao, C Xiao… - Talanta, 2013 - Elsevier</td>
#>       <td class="gt_row gt_left">A novel high performance liquid chromatography-electrochemical detector (HPLC-ECD) analytical system was developed in this study by integratedly utilizing ionic liquid (IL) of 1-butyl-3-methylimidazolium bromide and an additive of gold nanoparticles. The resulted pilot …</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">Intravenous iron sucrose in peritoneal dialysis patients with renal anemia</td>
#>       <td class="gt_row gt_left">http://www.pdiconnect.com/content/28/2/149.short</td>
#>       <td class="gt_row gt_left">www.pdiconnect.com</td>
#>       <td class="gt_row gt_left">H Li, SX Wang - Peritoneal dialysis international, 2008 - pdiconnect.com</td>
#>       <td class="gt_row gt_left">♦ Objective: To explore the safety and efficacy of intravenous (IV) iron sucrose in maintenance peritoneal dialysis (PD).♦ Design: Randomized, controlled, parallel-group single-center trial.♦ Setting: Blood Purification Center of Chaoyang, Beijing Capital …</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">Copper Nanoparticles Embedded in Lithium Tantalate Crystals for Multi-GHz Lasers</td>
#>       <td class="gt_row gt_left">https://pubs.acs.org/doi/abs/10.1021/acsanm.9b01321</td>
#>       <td class="gt_row gt_left">pubs.acs.org</td>
#>       <td class="gt_row gt_left">…, H Amekura, S Wang, H Yu, J Wang… - ACS Applied Nano …, 2019 - ACS Publications</td>
#>       <td class="gt_row gt_left">Embedded copper nanoparticles (Cu NPs) with spherical geometry are synthesized by Cu+ ion implantation in lithium tantalate (LiTaO3) crystal. Because of the localized surface plasmon resonance (LSPR) effect, strong optical absorption in the visible light band has …</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">Sex differences in cancer immunotherapy efficacy, biomarkers, and therapeutic strategy</td>
#>       <td class="gt_row gt_left">https://www.mdpi.com/527894</td>
#>       <td class="gt_row gt_left">www.mdpi.com</td>
#>       <td class="gt_row gt_left">S Wang, LA Cowley, XS Liu - Molecules, 2019 - mdpi.com</td>
#>       <td class="gt_row gt_left">Sex differences in innate and adaptive immune responses are known, and women generally mount a stronger immune response than men. Cancer immunotherapy, represented by immune checkpoint inhibitors (ICIs), has revolutionized the treatment of cancer, and sex …</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">[HTML][HTML] High neutrophil-to-lymphocyte ratio predicts cardiovascular mortality in chronic hemodialysis patients</td>
#>       <td class="gt_row gt_left">https://www.hindawi.com/journals/mi/2017/9327136/abs/</td>
#>       <td class="gt_row gt_left">www.hindawi.com</td>
#>       <td class="gt_row gt_left">H Li, X Lu, R Xiong, S Wang - Mediators of Inflammation, 2017 - hindawi.com</td>
#>       <td class="gt_row gt_left">The neutrophil-to-lymphocyte ratio (NLR) is a novel simple biomarker of inflammation. It has emerged as a predictor of poor prognosis in cancer and cardiovascular disease in general population. But little was known of its prognostic value in chronic hemodialysis (HD) …</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">Intravenous iron sucrose in Chinese hemodialysis patients with renal anemia</td>
#>       <td class="gt_row gt_left">https://www.karger.com/Article/PDF/113529</td>
#>       <td class="gt_row gt_left">www.karger.com</td>
#>       <td class="gt_row gt_left">H Li, S Wang - Blood purification, 2008 - karger.com</td>
#>       <td class="gt_row gt_left">Background/Aims: Renal anemia is one of the commonest complications of chronic renal failure. Iron deficiency is the most common factor which affects the efficacy of recombinant human erythropoietin (EPO) therapy. Intravenous (iv) iron preparations are commonly used …</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">β2-Adrenoceptor affinity chromatography and its application in the screening of the active compounds from Semen Armeniacae Amarum</td>
#>       <td class="gt_row gt_left">https://link.springer.com/article/10.1007/s11434-007-0510-8</td>
#>       <td class="gt_row gt_left">link.springer.com</td>
#>       <td class="gt_row gt_left">XH Zheng, XF Zhao, R Yang, SX Wang, YM Wei… - Chinese Science …, 2008 - Springer</td>
#>       <td class="gt_row gt_left">Abstract β 2-Adrenoceptor (β 2-AR) was purified from the rabbit lung tissue by sepharose-salbutamol affinity chromatographic column. To prepare the β 2-AR stationary phase, β 2-AR was evenly immobilized on the surface of macro-pore silica with a mild chemical coupling …</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">High neutrophil-to-lymphocyte ratio is a significant predictor of cardiovascular and all-cause mortality in patients undergoing peritoneal dialysis</td>
#>       <td class="gt_row gt_left">https://www.karger.com/Article/Abstract/488696</td>
#>       <td class="gt_row gt_left">www.karger.com</td>
#>       <td class="gt_row gt_left">X Lu, S Wang, G Zhang, R Xiong, H Li - Kidney and Blood Pressure …, 2018 - karger.com</td>
#>       <td class="gt_row gt_left">Background/Aims: Chronic inflammation is associated with increased risk of cardiovascular death in patients with end-stage renal disease (ESRD). Although elevated neutrophil-to-lymphocyte ratio (NLR), a novel inflammatory marker, has been shown to predict …</td>
#>     </tr>
#>   </tbody>
#>   
#>   
#> </table></div>

# }